requirements.txt
audioop-lts==0.2.2
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
pydub==0.25.1
PyMySQL==1.1.2
PyPDF2==3.0.1
python-dotenv==1.2.2
SpeechRecognition==3.15.2
SQLAlchemy==2.0.48
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.15.0
Werkzeug==3.1.7
WTForms==3.2.1
